backend/requirements.txt
# runtime
Flask>=2.3,<4
Flask-Cors>=4.0
flask-jwt-extended>=4.6
python-dotenv>=1.0
psycopg2-binary>=2.9

# dev / test
pytest>=7.0
pytest-cov>=4.0
